separate swap from transfer views, add respective paths
