+++ src/portal/lib/transfer.ts
@@ -1,5 +1,5 @@
 import { web3FromAddress } from "@polkadot/extension-dapp"
-
+import { Keyring } from '@polkadot/keyring'
 export async function getDepositFee (api:any) {
     const fee = await api.query.tftBridgeModule.depositFee()
     return fee.toNumber() / 1e7
@@ -17,5 +17,15 @@ export async function getDepositFee (api:any) {
         .signAndSend(address, { signer: injector.signer }, callback)
     } catch (error) {
       console.log(`err while trying to get injector ${error}`)
+    }
+  }
+  export function checkAddress (address: string) {
+    const keyring = new Keyring({ type: 'sr25519' })
+    try {
+      keyring.addFromAddress(address)
+      return true
+    } catch (error) {
+      console.log(error)
+      return false
     }
   }
